requirements.txt
absl-py==0.5.0
aenum==2.1.2
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
chumpy==0.67.6
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
dask==0.19.4
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
networkx==2.2
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pbr==4.3.0
Pillow==5.3.0
plyfile==0.6
protobuf==3.6.1
pymp-pypi==0.4.2
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
subprocess32==3.5.3
tensorboard==1.6.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toolz==0.9.0
tqdm==4.26.0
Werkzeug==0.14.1
